dom/websocket/tests/test_websocket_sharedWorker.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 1090183</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<script class="testbody" type="text/javascript">

var sw = new SharedWorker('websocket_sharedWorker.js');
sw.port.onmessage = function(event) {
  if (event.data.type == 'finish') {
    SimpleTest.finish();
  } else if (event.data.type == 'status') {
    ok(event.data.status, event.data.msg);
  }
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
